Support phprelease/filelist section for Pyrus (validation)
Onion generates package.xml files specifing version="2.0" and using the
XSD of package-2.0. The problem is that the install-as attribute used by
Onion in file elements is not allowed in this version and Pyrus, which
performs XML validation using the XSD, refuses to install our packages
lamenting that the XML in package.xml is malformed.

Instead of the install-as attribute in file elements, it would be
probably better to use the phprelease node e.g.:

    <phprelease>
        <filelist>
            <install name="Predis/Client.php" as="Predis/Client.php" />
            <!-- rest of the list -->
        <filelist>
    </phprelease>